cloud-deployment/requirements.txt
tensorflow>=2.9.0
fastapi>=0.85.0
uvicorn>=0.18.3
pillow>=9.0.0
numpy>=1.22.0
python-multipart>=0.0.5
